requirements.txt
pystray>=0.19.4
wmi>=1.5.1
Pillow>=10.0.0